Implemented login/register, massive changes.
Added new module 'src/mud/password_hasher.py' to asynchronously hash passwords in another python process due to it being a CPU bound operation. THis needs the running event loop so it is implemented as a class, and being initialized in the server's main function. A reference to this initialized class is now attached to the Manager.

Added new module 'src/mud/database.py' to implement async database access functions. It needs a running event loop so it is inside a class that is initialized inside the server's main function, a reference to this now resides in the Manager.

Reworked how logging in and registering works. Removed the old registration from the Command interpreter and made the player reference a REQUIRED argument.

New registration and login module at 'src/mud/login_register.py' to handle database and password hashing and checking from login/registration attempts.

New connections registered at the Manager will now enter into a 5 attempt loop to either register or login in, else they will be forcefully disconnected.