Updated password hashing and login/register functions.
Updated 'src/password_hasher.py' to now take in the password and 'credentials' stored in the player character documents as inputs for hashing.

Updated 'src/login_register.py' to use now use the new database functions, pass the new 'credentials' in to the password hasher. Now registering is setting the 'location' of a player in database to [0,0,0]