Experimented with hooking into database change stream for chat loggin.
Successfully tried out hooking into the change stream of the database to pull filtered changes.

In this example, we push new elements onto an array field, limiting the size of it during the operation to the X last elements added, thus creating a log.

We watch for that change and return only newly added element under the same field name 'log'.