Changes to Manager and moved to its own package.
Player's command interpreter will now be started and watched for in the Manager, automatically ending the task once the last connection drops out of a player.

Manager now has a new initializer to initialize event loop specific classes.

Command interpreter no longer holds a reference to player, 'interpreter.process' now requires it to be passed in as an argument.