Massive changes to the Manager. Updated Command Interpreter.
Changed Manager to actually be a.. manager holding managers and not a sphagetti.

New ConnectionManager that will house all references to connections bound to players, providing an aggregate read/write queue, and handling disconnects. Also holds methods for sending and recieving messages from players, and watching for commands.

New AuthenticationManager that holds methods for registration/login of the players.

New PlayerManager that holds data on logged in players and their location, also implementing methods to manipulate and get player data. Also implements ways to recall all players in a room.

Changed commands/Interpreter to now holds a reference to it's own database client, a passed in reference to the main Manager object. Command functions now have the manager and player name passed in alongside the database and message.

Commands have been updated to use the new Manager interface.

New study: how to avoid circular import errors when I try to get the type hints done.

Various little changes and improvements across the board.